+++ packages/perseus-linter/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/perseus-linter
 
+## 0.3.6
+
+### Patch Changes
+
+-   Updated dependencies [22a9c408]
+    -   @khanacademy/perseus-core@1.1.2
+    -   @khanacademy/perseus-error@0.2.5
+
+## 0.3.5
+
+### Patch Changes
+
+-   55d4cd00: Print package name and version when loaded in the page
+-   Updated dependencies [55d4cd00]
+    -   @khanacademy/perseus-core@1.1.1
+    -   @khanacademy/perseus-error@0.2.4
+
 ## 0.3.4
 
 ### Patch Changes